refactor: move UNKNOWN_STRING to constants.general
