+++ backend/kernelCI_app/constants/general.py
@@ -1,1 +1,3 @@
 DEFAULT_ORIGIN = 'maestro'
+
+UNKNOWN_STRING = 'Unknown'
